correctness/requirements.txt
torch>=1.10.0
faiss-gpu>=1.7.0; sys_platform != 'darwin'
faiss-cpu>=1.7.0; sys_platform == 'darwin'
sentence-transformers>=2.2.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
openai>=1.0.0
psutil>=5.9.0
tqdm>=4.62.0
joblib>=1.1.0
scipy>=1.7.0
